open project src in editor through cmd opengoenv

+++ goenv.bash
@@ -3,40 +3,69 @@ lsgoenv(){
     ls -1t ~/.goenv/projects/
 }
 
+__seteditor(){
+    echo "Enter your preferred editor:"
+    read editor
+    if ! [ -z "$editor" ]; then
+        echo "\n EDITOR='$editor'" >> ~/.goenv/projects/$1
+    fi
+}
+
 mkgoenv(){
     if [ -z "$1" ]; then
-        echo "Add a valid project name."
+        echo "Enter a valid project name ."
         echo "eg.   mkgoenv project-name"
     else
         if [ -f ~/.goenv/projects/$1 ]; then
-            echo "$1 already exists."
+            source ~/.goenv/projects/$1
+            echo "Project $1 exists at '$PROJECT_PATH'."
         else
             echo "PROJECT_PATH='$PWD/$1'" > ~/.goenv/projects/$1
             echo "Creating project '$1' at '$PWD/'."
-            mkdir -p $1 && cd $1
-            mkdir -p bin
-            mkdir -p pkg
-            mkdir -p src
-            mkdir -p src/$1
-            export GOPATH=$(pwd)
-            cd src/$1
+            mkdir $1
+            mkdir $1/bin
+            mkdir $1/pkg
+            mkdir $1/src
+            mkdir $1/src/$1
+            touch $1/src/$1/main.go
+            export GOENV="$1"
+            export GOPATH="$PWD/$GOENV"
+            __seteditor $1
+            if ! [ -z "$editor" ]; then
+                cd $1/src/$1
+                $editor .
+            fi
         fi
     fi
 }
 
 rmgoenv(){
     if [ -z "$1" ]; then
-        echo "no valid project name entered."
+        echo "Enter a valid project name."
         echo "eg.   rmgoenv project-name"
         lsgoenv
     else
         if [ -f ~/.goenv/projects/$1 ]; then
             source ~/.goenv/projects/$1
-            echo "Deleting project '$PROJECT_PATH'"
-            rm -rf $PROJECT_PATH/
+            echo "Do you wish to delete project files from drive [y/n]? (n)"
+            read yn
+            case $yn in
+                [Yy] )
+                    source ~/.goenv/projects/$1
+                    echo "Deleting project '$PROJECT_PATH'."
+                    rm -rf $PROJECT_PATH/
+                    ;;
+                * ) echo "Removing goenv '$1'."
+                    echo "Project files still at '$PROJECT_PATH'."
+                    ;;
+            esac
             rm -f ~/.goenv/projects/$1
+            if [ $1=$GOENV ]; then
+                unset GOENV
+                unset GOPATH
+            fi
         else
-            echo "project '$1' not found."
+            echo "'$1' not found."
             lsgoenv
         fi
     fi
@@ -50,19 +79,20 @@ goenv(){
     else
         if [ -f ~/.goenv/projects/$1 ]; then
             source ~/.goenv/projects/$1
-            echo "Loading project '$PROJECT_PATH'"
+            echo "Activating goenv for '$PROJECT_PATH'"
+            export GOENV=$1
             export GOPATH=$PROJECT_PATH/
-            cd $GOPATH/
         else
-            echo "Project '$1' not found."
+            echo "'$1' not found."
             lsgoenv
         fi
     fi
 }
 
 dropgoenv(){
-    if ! [ -z $GOPATH ]; then
-        echo "dropping GOPATH: $GOPATH"
+    if ! [ -z $GOENV ]; then
+        echo "Dropping GOENV: '$GOENV'"
+        unset GOENV
         unset GOPATH
     fi
 }
@@ -72,15 +102,33 @@ addtogoenv(){
         echo "Enter in given format."
         echo "eg.   addtogoenv project-name project-path(GOPATH)"
     else
-        echo "Adding project '$2' to goenv."
-        echo "PROJECT_PATH='$2'" > ~/.goenv/projects/$1
+        if [ -f ~/.goenv/projects/$1 ]; then
+            source ~/.goenv/projects/$1
+            echo "goenv '$1' exists."
+            echo "Project exists at '$PROJECT_PATH'."
+        else
+            echo "Adding project '$2' to goenv '$1'."
+            echo "PROJECT_PATH='$2'" > ~/.goenv/projects/$1
+            __seteditor $1
+        fi
     fi
 }
 
 whichgoenv(){
-    if [ -z "$GOPATH" ]; then
+    if [ -z "$GOENV" ]; then
         echo "No goenv set."
     else
-        echo "GOPATH=$GOPATH"
+        echo "GOENV: $GOENV"
+        echo "GOPATH: $GOPATH"
+    fi
+}
+
+opengoenv(){
+    goenv $1
+    if ! [ -z "$1" ]; then
+        cd $GOPATH/src/$1
+        if ! [ -z $EDITOR ]; then
+            $EDITOR .
+        fi
     fi
 }
